requirements.txt
Flask
flask-pymongo
werkzeug
qrcode
Pillow
reportlab
